refactor: 음식점 목록 정렬/렌더링하는 로직 useRestaurantList 커스텀 훅으로 이동
