+++ src/components/Main.tsx
@@ -1,12 +1,11 @@
 import { useEffect } from 'react';
 import { saveToLocalStorage } from '../utils/localStorage';
 import { getRestaurantListData } from '../data/restaurantListData';
-import { filterAndSortRestaurantList } from '../domains/restaurantUtil';
 import FilterSection from './FilterSection/FilterSection';
 import RestaurantList from './RestaurantList/RestaurantList';
 import Modal from './common/Modal/Modal';
 import RestaurantDetail from './RestaurantDetail/RestaurantDetail';
-import { useRestaurantList } from '../hooks/useRestaurantList';
+import { filterAndSortRestaurantList, useRestaurantList } from '../hooks/useRestaurantList';
 import { useSelectRestaurant } from '../hooks/useSelectRestaurant';
 
 const restaurantList = filterAndSortRestaurantList(getRestaurantListData());
